requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.12.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.13.0
packaging==24.0
pandas==2.2.3
pillow==10.4.0
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==18.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.19.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.47
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.18.1
